requirements.txt
fastapi
uvicorn
websockets
twilio
deepgram-sdk
fastapi
uvicorn
websockets
twilio
deepgram-sdk
google-genai
httpx
python-dotenv
pydantic
pyjwt
passlib
bcrypt==3.2.2
python-multipart
chromadb
pymupdf
numpy
groq
faiss-cpu
sentence-transformers
